server: split containers actions
