+++ server/utils.go
@@ -5,6 +5,8 @@ import (
 	"io"
 	"os"
 	"strings"
+
+	"github.com/kubernetes-incubator/cri-o/oci"
 )
 
 const (
@@ -12,6 +14,28 @@ const (
 	// "The search list is currently limited to six domains with a total of 256 characters."
 	maxDNSSearches = 6
 )
+
+type containerRequest interface {
+	GetContainerId() string
+}
+
+func (s *Server) getContainerFromRequest(req containerRequest) (*oci.Container, error) {
+	ctrID := req.GetContainerId()
+	if ctrID == "" {
+		return nil, fmt.Errorf("container ID should not be empty")
+	}
+
+	containerID, err := s.ctrIDIndex.Get(ctrID)
+	if err != nil {
+		return nil, fmt.Errorf("container with ID starting with %s not found: %v", ctrID, err)
+	}
+
+	c := s.state.containers.Get(containerID)
+	if c == nil {
+		return nil, fmt.Errorf("specified container not found: %s", containerID)
+	}
+	return c, nil
+}
 
 func int64Ptr(i int64) *int64 {
 	return &i
